src/backend_v2/requirements.txt
flask
flask-cors
pandas
openpyxl
flask_sqlalchemy
python-dotenv
psycopg2-binary==2.9.10